vm keeper refactor
