+++ examples/gno.land/r/demo/hello/hello_filetest.gnoa
