+++ examples/gno.land/r/demo/hello/msh_filetest.gnoa
@@ -1,0 +1,27 @@
+package main 
+
+import "std"
+
+// type Proposal struct {
+// 	TextContent string `json:"text_content,omitempty"`
+// }
+
+
+func main(){
+	call := &std.CallMsg{
+		PkgPath: "gno.land/r/demo/greet",
+		Fn: "Greet",     
+		Args: []string{"hey"}, 
+	}
+
+	// prop := &Proposal{TextContent: "hey"}
+
+	r, ok := std.Marshal(call)	
+	if ok{
+		println(string(r))
+	}
+	println("done")
+}
+
+// Output:
+// done
