+++ gno.land/keeper/common_test.go
@@ -1,4 +1,4 @@
-package vm
+package vmk
 
 // DONTCOVER
