+++ gno.land/keeper/keeper_test.goa
@@ -1,4 +1,4 @@
-package vm
+package vmk
 
 // TODO: move most of the logic in ROOT/gno.land/...
