+++ gnovm/stdlibs/channel.goa
@@ -1,10 +1,11 @@
 package stdlibs
 
 type CallMsg struct {
-	Caller  string   `json:"caller,omitempty"`
-	PkgPath string   `json:"pkg_path,omitempty"`
-	Fn      string   `json:"fn,omitempty"`
-	Args    []string `json:"args,omitempty"`
+	ChainID string
+	Caller  string
+	PkgPath string
+	Fn      string
+	Args    []string
 }
 
 type Request struct {
@@ -12,23 +13,28 @@ type Request struct {
 	Callback *CallMsg
 }
 
-var MsgQueue chan *Request
+var msgQueue chan *Request
 
 // deprecated
 // var ResQueue chan string // receive response with timeout, used for synchronous call
 
 func init() {
-	MsgQueue = make(chan *Request)
+	msgQueue = make(chan *Request)
 	// ResQueue = make(chan string)
 }
 
 // this is called by stdlib function
 func Send(call, callback *CallMsg, mc chan<- *Request) {
+	println("send request")
 	req := &Request{
 		Call:     call,
 		Callback: callback,
 	}
 	mc <- req
+}
+
+func GetInternalQueue() <-chan *Request {
+	return msgQueue
 }
 
 // synchronous call, use ResQueue to get result
