+++ gnovm/stdlibs/std/msg.gno
@@ -1,10 +1,10 @@
 package std
 
 // TODO: a better encoder
-func EncodeMsgCall(PkgPath, Fn string, args []string) string {
+func EncodeMsgCall(ChainID, PkgPath, Fn string, args []string) string {
 	var as string
 	for _, a := range args {
 		as += a
 	}
-	return PkgPath+"#"+Fn+"#"+as
+	return ChainID + "#" + PkgPath + "#" + Fn + "#" + as
 }
